Fix: #65
When adding more than one disallowed content type alias, the query for them went wrong, excluding all nodes with the content type aliases in the content. In the report, we had a site, where the content type alias "person" was disallowed, and all nodes with the word "person" in FullTextContent got excluded. This is fixed by setting the field name on all queries.